🧹 Extract INMET data fetch logic into a reusable helper
