+++ pages/inmet.js
@@ -2,6 +2,7 @@ import React, { useState, useEffect } from 'react';
 import Head from 'next/head';
 import { useI18n } from '../lib/i18n/i18nContext';
 import styles from '../styles/Inmet.module.css';
+import { fetchInmetData } from '../lib/api/inmet';
 
 export default function InmetCapitals() {
   const { t } = useI18n();
@@ -12,45 +13,9 @@ export default function InmetCapitals() {
   useEffect(() => {
     const fetchData = async () => {
       try {
-        const today = new Date();
-        const year = today.getFullYear();
-        const month = String(today.getMonth() + 1).padStart(2, '0');
-        const day = String(today.getDate()).padStart(2, '0');
-        const dateStr = `${year}-${month}-${day}`;
-
-        const res = await fetch(`https://apitempo.inmet.gov.br/condicao/capitais/${dateStr}`);
-        if (!res.ok) {
-          throw new Error(`HTTP error! status: ${res.status}`);
-        }
-        const json = await res.json();
-
-        // Handle INMET returning an object instead of array sometimes, or an empty response
-        if (Array.isArray(json)) {
-            setData(json);
-        } else if (json && Object.keys(json).length > 0) {
-            // sometimes it returns an object of objects
-            setData(Object.values(json));
-        } else {
-             // Try yesterday if today fails or returns empty
-             const yesterday = new Date(today);
-             yesterday.setDate(yesterday.getDate() - 1);
-             const yYear = yesterday.getFullYear();
-             const yMonth = String(yesterday.getMonth() + 1).padStart(2, '0');
-             const yDay = String(yesterday.getDate()).padStart(2, '0');
-             const yDateStr = `${yYear}-${yMonth}-${yDay}`;
-
-             const yRes = await fetch(`https://apitempo.inmet.gov.br/condicao/capitais/${yDateStr}`);
-             if (yRes.ok) {
-                 const yJson = await yRes.json();
-                 if (Array.isArray(yJson)) {
-                     setData(yJson);
-                 } else if (yJson && Object.keys(yJson).length > 0) {
-                     setData(Object.values(yJson));
-                 }
-             } else {
-                 throw new Error("No data available for today or yesterday");
-             }
-        }
+        setLoading(true);
+        const fetchedData = await fetchInmetData();
+        setData(fetchedData);
       } catch (e) {
         console.error("Failed to fetch INMET data", e);
         setError(e.message);
